backend/requirements.txt
fastapi
uvicorn
pymongo
bcrypt
python-dotenv
pydantic[email]
cryptography